move sample categories panel to owned folder
