+++ src/app/app.module.ts
@@ -23,7 +23,7 @@ import { MethodBadgeComponent } from './method-badge.component';
 import { QueryRowComponent } from './queryrow/queryrow.component';
 import { RequestEditorsComponent } from './request-editors/request-editors.component';
 import { ResponseStatusBarComponent } from './response-status-bar/response-status-bar.component';
-import { SampleCategoriesPanelComponent } from './sample-categories-panel.component';
+import { SampleCategoriesPanelComponent } from './sample-categories/sample-categories-panel.component';
 import { ScopesDialogComponent } from './scopes-dialog/scopes-dialog.component';
 import { ShareLinkBtnComponent } from './share-link/share-link-btn.component';
 import { SidebarComponent } from './sidebar/sidebar.component';
