+++ src/app/sample-categories/sample-categories-panel.component.html
